Move field stuff to separate source file
