+++ core/common/src/annotation/TextBlock.ts
@@ -8,6 +8,7 @@
 
 import { Id64String } from "@itwin/core-bentley";
 import { TextStyleSettings, TextStyleSettingsProps } from "./TextStyle";
+import { FieldFormatter, FieldPropertyHost, FieldPropertyPath } from "./TextField";
 
 /** Options supplied to [[TextBlockComponent.clearStyleOverrides]] to control how the style overrides are cleared on the component and its child components.
  * @beta
@@ -373,57 +374,6 @@ export class TabRun extends TextBlockComponent {
     return other instanceof TabRun && super.equals(other);
   }
 }
-
-/** A chain of property accesses that resolves to a primitive value that forms the basis of the displayed content
- * of a [[FieldRun]].
-   * The simplest property paths consist of a [[propertyName]] and nothing else, where `propertyName` identifies
-   * a primitive property.
-   * If `propertyName` identifies a struct or array property, then additional [[accessors]] are required to identify the specific value.
-   * If `propertyName` (including any [[accessors]]) resolves to a JSON property, then additional [[jsonAccessors]] are required to identify a specific value within the JSON.
-   * Some examples:
-   * ```
-   * | Access String | propertyName | accessors | jsonAccessors |
-   * | ------------- | ------------ | --------- | ------------- |
-   * | name          | "name"       | undefined | undefined     |
-   * | spouse.name   | "spouse"     | [name]    | undefined     |
-   * | colors[2]     | "colors"     | [2]       | undefined     |
-   * | spouse.favoriteRestaurants[1].address | "spouse" | ["favoriteRestaurants", 1, "address"] | undefined |
-   * | jsonProperties.contactInfo.email | "jsonProperties" | undefined | ["contactInfo", "email"] |
-   * | spouse.jsonProperties.contactInfo.phoneNumbers[0].areaCode | "spouse" | ["jsonProperties"] | ["contactInfo", "phoneNumbers", 0, "areaCode"] |
-   * ```
- * @beta
- */
-export interface FieldPropertyPath {
-  /** The name of the BIS property of the [[FieldPropertyHost]] that serves as the root of the path. */
-  propertyName: string;
-  /** Property names and/or array indices describing the path from [[propertyName]] to the ultimate BIS property. */
-  accessors?: Array<string | number>;
-  /** If [[propertyName]] and [[accessors]] (if defined) resolve to a BIS property of extended type `Json`, property names and/or
-   * array indices for selecting a primitive value within the JSON.
-   */
-  jsonAccessors?: Array<string | number>;
-}
-
-/** Describes the source of the property value against which a [[FieldPropertyPath]] is evaluated.
- * A field property is always hosted by an [Element]($backend). It may be a property of the element's BIS class itself,
- * or that of one of its [ElementAspect]($backend)s.
- * The [[schemaName]] and [[className]] should always identify the exact class that contains [[FieldPropertyPath.propertyName]] - not a subclass thereof.
- * @beta
- */
-export interface FieldPropertyHost {
-  /** The Id of the [Element]($backend) that hosts the property. */
-  elementId: Id64String;
-  /** The name of the schema containing the class identified by [[className]]. */
-  schemaName: string;
-  /** The name of the exact class (not a subclass) containing the property identified by [[FieldPropertyPath.propertyName]]. */
-  className: string;
-}
-
-/** Placeholder type for a description of how to format the raw property value resolved by a [[FieldPropertyPath]] into a [[FieldRun]]'s display string.
- * *** COMING SOON ***
- * @beta
- */
-export interface FieldFormatter { [k: string]: any }
 
 /** JSON representation of a [[FieldRun]].
  * @beta
