+++ core/common/src/core-common.ts
@@ -8,6 +8,7 @@ export * from "./annotation/TextAnnotation";
 export * from "./annotation/TextBlock";
 export * from "./annotation/TextBlockGeometryProps";
 export * from "./annotation/TextBlockLayoutResult";
+export * from "./annotation/TextField";
 export * from "./annotation/TextStyle";
 export * from "./Atmosphere";
 export * from "./AuthorizationClient";
